feat(codecompanion)!: migrate cc extension to use codecompanion.extensions api
BREAKING:
codecompanion extension is now more streamlined in one place.

```lua
require("codecompanion").setup({
--other config
extensions = {
  mcphub = {
    callback = "mcphub.extensions.codecompanion",
    opts = {
      make_vars = true,
      make_slash_commands = true,
      show_result_in_chat = true,
    },
  },
 },
})
```

* The `@mcp` tool is split into two separate tools `use_mcp_tool` and
`access_mcp_resource`.
* IMPORTANT: remove old mcp tool declared in
codecompanion.config.strategies.chat.tools
* mcphub.config.extensions.codecompanion is deprecated as the options
are now declared at codecompanion extension itself.
