+++ lua/mcphub/extensions/codecompanion/utils.lua
@@ -1,10 +1,7 @@
 local M = {}
-local State = require("mcphub.state")
-local config = require("codecompanion.config")
-local log = require("mcphub.utils.log")
 local shared = require("mcphub.extensions.shared")
 
-function M.create_handler(action_name, has_function_calling)
+function M.create_handler(action_name, has_function_calling, opts)
     return function(agent, args, _, output_handler)
         local params = shared.parse_params(args, action_name)
         if #params.errors > 0 then
@@ -80,8 +77,9 @@ local function replace_headers(text)
     end
     return table.concat(lines, "\n")
 end
-local function add_tool_output(action_name, tool, chat, llm_msg, is_error, has_function_calling)
-    local show_result_in_chat = State.config.extensions.codecompanion.show_result_in_chat == true
+local function add_tool_output(action_name, tool, chat, llm_msg, is_error, has_function_calling, opts)
+    local config = require("codecompanion.config")
+    local show_result_in_chat = opts.show_result_in_chat == true
     local text = show_result_in_chat and replace_headers(llm_msg) or llm_msg
     if has_function_calling then
         chat:add_tool_output(tool, text, (show_result_in_chat or is_error) and text or "Tool result shared")
@@ -104,7 +102,7 @@ local function add_tool_output(action_name, tool, chat, llm_msg, is_error, has_f
     end
 end
 
-function M.create_output_handlers(action_name, has_function_calling)
+function M.create_output_handlers(action_name, has_function_calling, opts)
     return {
         error = function(self, agent, cmd, stderr)
             local stderr = has_function_calling and (stderr[1] or "") or cmd[1]
@@ -121,12 +119,13 @@ function M.create_output_handlers(action_name, has_function_calling)
                 action_name,
                 stderr
             )
-            add_tool_output(action_name, self, agent.chat, err_msg, true, has_function_calling)
+            add_tool_output(action_name, self, agent.chat, err_msg, true, has_function_calling, opts)
         end,
         success = function(self, agent, cmd, stdout)
             local result = has_function_calling and stdout[1] or cmd[1]
             agent = has_function_calling and agent or self
             -- Show text content if present
+            -- TODO: add messages with role = `tool` when supported
             if result.text and result.text ~= "" then
                 local to_llm = string.format(
                     [[The `%s` call returned the following text:
@@ -134,11 +133,192 @@ function M.create_output_handlers(action_name, has_function_calling)
                     action_name,
                     result.text
                 )
-                add_tool_output(action_name, self, agent.chat, to_llm, false, has_function_calling)
+                add_tool_output(action_name, self, agent.chat, to_llm, false, has_function_calling, opts)
             end
             -- TODO: Add image support when codecompanion supports it
         end,
     }
+end
+
+function M.setup_codecompanion_variables(opts)
+    if not opts.make_vars then
+        return
+    end
+    local mcphub = require("mcphub")
+    --setup event listners to update variables, tools etc
+    mcphub.on({ "servers_updated", "resource_list_changed" }, function(_)
+        local hub = mcphub.get_hub_instance()
+        if not hub then
+            return
+        end
+        local resources = hub:get_resources()
+        local ok, config = pcall(require, "codecompanion.config")
+        if not ok then
+            return
+        end
+
+        local cc_variables = config.strategies.chat.variables
+        -- remove existing mcp variables that start with mcp
+        for key, value in pairs(cc_variables) do
+            local id = value.id or ""
+            if id:sub(1, 3) == "mcp" then
+                cc_variables[key] = nil
+            end
+        end
+        for _, resource in ipairs(resources) do
+            local server_name = resource.server_name
+            local uri = resource.uri
+            local resource_name = resource.name or uri
+            local description = resource.description or ""
+            if type(description) == "function" then
+                local ok, desc = pcall(description, resource)
+                if ok then
+                    description = desc or ""
+                else
+                    description = "Error in description function: " .. (desc or "")
+                end
+            end
+            --remove new lines
+            description = description:gsub("\n", " ")
+
+            description = resource_name .. " (" .. description .. ")"
+            cc_variables[uri] = {
+                id = "mcp" .. server_name .. uri,
+                description = description,
+                callback = function(self)
+                    -- this is sync and will block the UI (can't use async in variables yet)
+                    local response = hub:access_resource(server_name, uri, {
+                        caller = {
+                            type = "codecompanion",
+                            codecompanion = self,
+                            meta = {
+                                is_within_variable = true,
+                            },
+                        },
+                        parse_response = true,
+                    })
+                    return response and response.text
+                end,
+            }
+        end
+    end)
+end
+
+function M.setup_codecompanion_slash_commands(opts)
+    if not opts.make_slash_commands then
+        return
+    end
+
+    local mcphub = require("mcphub")
+    local config = require("codecompanion.config")
+    --setup event listners to update variables, tools etc
+    mcphub.on({ "servers_updated", "prompt_list_changed" }, function(_)
+        local hub = mcphub.get_hub_instance()
+        if not hub then
+            return
+        end
+        local prompts = hub:get_prompts()
+        local slash_commands = config.strategies.chat.slash_commands
+        -- remove existing mcp slash commands that start with mcp
+        for key, value in pairs(slash_commands) do
+            local id = value.id or ""
+            if id:sub(1, 3) == "mcp" then
+                slash_commands[key] = nil
+            end
+        end
+        for _, prompt in ipairs(prompts) do
+            local server_name = prompt.server_name
+            local prompt_name = prompt.name or ""
+            local description = prompt.description or ""
+            local arguments = prompt.arguments or {}
+            if type(description) == "function" then
+                local ok, desc = pcall(description, prompt)
+                if ok then
+                    description = desc or ""
+                else
+                    description = "Error in description function: " .. (desc or "")
+                end
+            end
+            if type(arguments) == "function" then
+                local ok, args = pcall(arguments, prompt)
+                if ok then
+                    arguments = args or {}
+                else
+                    vim.notify("Error in arguments function: " .. (args or ""), vim.log.levels.ERROR)
+                    arguments = {}
+                end
+            end
+            --remove new lines
+            description = description:gsub("\n", " ")
+
+            description = prompt_name .. " (" .. description .. ")"
+            slash_commands["mcp:" .. prompt_name] = {
+                id = "mcp" .. server_name .. prompt_name,
+                description = description,
+                callback = function(self)
+                    shared.collect_arguments(arguments, function(values)
+                        -- this is sync and will block the UI (can't use async in slash_commands yet)
+                        local response, err = hub:get_prompt(server_name, prompt_name, values, {
+                            caller = {
+                                type = "codecompanion",
+                                codecompanion = self,
+                                meta = {
+                                    is_within_slash_command = true,
+                                },
+                            },
+                            parse_response = true,
+                        })
+                        if not response then
+                            if err then
+                                vim.notify("Error in slash command: " .. err, vim.log.levels.ERROR)
+                                vim.notify("Prompt cancelled", vim.log.levels.INFO)
+                            end
+                            return
+                        end
+                        local messages = response.messages or {}
+                        local text_messages = 0
+                        for i, message in ipairs(messages) do
+                            local output = message.output
+                            --TODO: Currently codecompanion only supports text messages
+                            if output.text and output.text ~= "" then
+                                local mapped_role = message.role == "assistant" and config.constants.LLM_ROLE
+                                    or message.role == "system" and config.constants.SYSTEM_ROLE
+                                    or config.constants.USER_ROLE
+                                text_messages = text_messages + 1
+                                -- if last message is from user, add it to the chat buffer
+                                if i == #messages and mapped_role == config.constants.USER_ROLE then
+                                    self:add_buf_message({
+                                        role = mapped_role,
+                                        content = output.text,
+                                    })
+                                else
+                                    self:add_message({
+                                        role = mapped_role,
+                                        content = output.text,
+                                    })
+                                end
+                            end
+                        end
+                        vim.notify(
+                            string.format(
+                                "%s message%s added successfully",
+                                text_messages,
+                                text_messages == 1 and "" or "s"
+                            ),
+                            vim.log.levels.INFO
+                        )
+                    end)
+                end,
+            }
+        end
+    end)
+end
+
+function M.setup_codecompanion_tools(enabled)
+    if not enabled then
+        return
+    end
+    --INFO:Individual tools might be an overkill
 end
 
 return M
