+++ lua/mcphub/extensions/shared.lua
@@ -43,191 +43,6 @@ function M.parse_params(params, action_name)
     }
 end
 
-function M.setup_codecompanion_variables(enabled)
-    if not enabled then
-        return
-    end
-    local mcphub = require("mcphub")
-    --setup event listners to update variables, tools etc
-    mcphub.on({ "servers_updated", "resource_list_changed" }, function(_)
-        local hub = mcphub.get_hub_instance()
-        if not hub then
-            return
-        end
-        local resources = hub:get_resources()
-        local ok, config = pcall(require, "codecompanion.config")
-        if not ok then
-            return
-        end
-
-        local cc_variables = config.strategies.chat.variables
-        -- remove existing mcp variables that start with mcp
-        for key, value in pairs(cc_variables) do
-            local id = value.id or ""
-            if id:sub(1, 3) == "mcp" then
-                cc_variables[key] = nil
-            end
-        end
-        for _, resource in ipairs(resources) do
-            local server_name = resource.server_name
-            local uri = resource.uri
-            local resource_name = resource.name or uri
-            local description = resource.description or ""
-            if type(description) == "function" then
-                local ok, desc = pcall(description, resource)
-                if ok then
-                    description = desc or ""
-                else
-                    description = "Error in description function: " .. (desc or "")
-                end
-            end
-            --remove new lines
-            description = description:gsub("\n", " ")
-
-            description = resource_name .. " (" .. description .. ")"
-            cc_variables[uri] = {
-                id = "mcp" .. server_name .. uri,
-                description = description,
-                callback = function(self)
-                    -- this is sync and will block the UI (can't use async in variables yet)
-                    local response = hub:access_resource(server_name, uri, {
-                        caller = {
-                            type = "codecompanion",
-                            codecompanion = self,
-                            meta = {
-                                is_within_variable = true,
-                            },
-                        },
-                        parse_response = true,
-                    })
-                    return response and response.text
-                end,
-            }
-        end
-    end)
-end
-
-function M.setup_codecompanion_slash_commands(enabled)
-    if not enabled then
-        return
-    end
-
-    local mcphub = require("mcphub")
-    --setup event listners to update variables, tools etc
-    mcphub.on({ "servers_updated", "prompt_list_changed" }, function(_)
-        local hub = mcphub.get_hub_instance()
-        if not hub then
-            return
-        end
-        local prompts = hub:get_prompts()
-        local ok, config = pcall(require, "codecompanion.config")
-        if not ok then
-            return
-        end
-
-        local slash_commands = config.strategies.chat.slash_commands
-        -- remove existing mcp slash commands that start with mcp
-        for key, value in pairs(slash_commands) do
-            local id = value.id or ""
-            if id:sub(1, 3) == "mcp" then
-                slash_commands[key] = nil
-            end
-        end
-        for _, prompt in ipairs(prompts) do
-            local server_name = prompt.server_name
-            local prompt_name = prompt.name or ""
-            local description = prompt.description or ""
-            local arguments = prompt.arguments or {}
-            if type(description) == "function" then
-                local ok, desc = pcall(description, prompt)
-                if ok then
-                    description = desc or ""
-                else
-                    description = "Error in description function: " .. (desc or "")
-                end
-            end
-            if type(arguments) == "function" then
-                local ok, args = pcall(arguments, prompt)
-                if ok then
-                    arguments = args or {}
-                else
-                    vim.notify("Error in arguments function: " .. (args or ""), vim.log.levels.ERROR)
-                    arguments = {}
-                end
-            end
-            --remove new lines
-            description = description:gsub("\n", " ")
-
-            description = prompt_name .. " (" .. description .. ")"
-            slash_commands["mcp:" .. prompt_name] = {
-                id = "mcp" .. server_name .. prompt_name,
-                description = description,
-                callback = function(self)
-                    M.collect_arguments(arguments, function(values)
-                        -- this is sync and will block the UI (can't use async in slash_commands yet)
-                        local response, err = hub:get_prompt(server_name, prompt_name, values, {
-                            caller = {
-                                type = "codecompanion",
-                                codecompanion = self,
-                                meta = {
-                                    is_within_slash_command = true,
-                                },
-                            },
-                            parse_response = true,
-                        })
-                        if not response then
-                            if err then
-                                vim.notify("Error in slash command: " .. err, vim.log.levels.ERROR)
-                                vim.notify("Prompt cancelled", vim.log.levels.INFO)
-                            end
-                            return
-                        end
-                        local messages = response.messages or {}
-                        local text_messages = 0
-                        for i, message in ipairs(messages) do
-                            local output = message.output
-                            --TODO: Currently codecompanion only supports text messages
-                            if output.text and output.text ~= "" then
-                                local mapped_role = message.role == "assistant" and config.constants.LLM_ROLE
-                                    or message.role == "system" and config.constants.SYSTEM_ROLE
-                                    or config.constants.USER_ROLE
-                                text_messages = text_messages + 1
-                                -- if last message is from user, add it to the chat buffer
-                                if i == #messages and mapped_role == config.constants.USER_ROLE then
-                                    self:add_buf_message({
-                                        role = mapped_role,
-                                        content = output.text,
-                                    })
-                                else
-                                    self:add_message({
-                                        role = mapped_role,
-                                        content = output.text,
-                                    })
-                                end
-                            end
-                        end
-                        vim.notify(
-                            string.format(
-                                "%s message%s added successfully",
-                                text_messages,
-                                text_messages == 1 and "" or "s"
-                            ),
-                            vim.log.levels.INFO
-                        )
-                    end)
-                end,
-            }
-        end
-    end)
-end
-
-function M.setup_codecompanion_tools(enabled)
-    if not enabled then
-        return
-    end
-    --INFO:Individual tools might be an overkill
-end
-
 function M.collect_arguments(arguments, callback)
     local values = {}
     local should_proceed = true
